Adds car status change functionality
Implements car status change functionality on the car details page, allowing users to set a car's status to "on event", "under maintenance" or "available".
This provides users a confirmation dialog before proceeding with the status change.
It also adjusts the fleet view to use `isPending` instead of `isLoading` from the `useCars` hook.